1.)remove unused methods 2.) adjust eventExecuteService position, need to be before register the eventcallback
