+++ modEvtMgrImpl/src/org/aion/evtmgr/impl/handler/ConsensusHandler.java
@@ -25,10 +25,8 @@
 package org.aion.evtmgr.impl.handler;
 
 import org.aion.evtmgr.IEvent;
-import org.aion.evtmgr.IEventCallback;
 import org.aion.evtmgr.IHandler;
 import org.aion.evtmgr.impl.abs.AbstractHandler;
-import org.aion.evtmgr.impl.callback.EventCallbackA0;
 
 /**
  * @author jay
@@ -40,34 +38,6 @@ public class ConsensusHandler extends AbstractHandler implements IHandler {
     public ConsensusHandler() {
         dispatcher.setName("ConsHdr");
     }
-
-    @SuppressWarnings("rawtypes")
-    public <E extends IEvent> void dispatch(E event) {
-        if (this.typeEqual(event.getEventType())) {
-
-            if (LOG.isTraceEnabled()) {
-                LOG.trace("CB size:[{}] cbType:[{}]", this.eventCallback.size(), event.getCallbackType());
-            }
-
-            for (IEventCallback cb : this.eventCallback) {
-                cb.onEvent(event);
-
-//                switch (event.getCallbackType()) {
-//                case 0:
-//                    ((EventCallbackA0) cb).onSyncDone();
-//                    break;
-//                case 1:
-//                    ((EventCallbackA0) cb).onBlockTemplate(event.getFuncArgs().get(0));
-//                    break;
-//                case 2:
-//                    ((EventCallbackA0) cb).onSolution(event.getFuncArgs().get(0));
-//                    break;
-//                default:
-//                }
-            }
-        }
-    }
-
     /*
      * (non-Javadoc)
      *
