+++ modEvtMgrImpl/src/org/aion/evtmgr/impl/handler/MinerHandler.java
@@ -25,10 +25,8 @@
 package org.aion.evtmgr.impl.handler;
 
 import org.aion.evtmgr.IEvent;
-import org.aion.evtmgr.IEventCallback;
 import org.aion.evtmgr.IHandler;
 import org.aion.evtmgr.impl.abs.AbstractHandler;
-import org.aion.evtmgr.impl.callback.EventCallbackA0;
 
 /**
  * @author jay
@@ -38,37 +36,6 @@ public class MinerHandler extends AbstractHandler implements IHandler {
 
     public MinerHandler() {
         dispatcher.setName("MinerHdr");
-    }
-
-    @SuppressWarnings("rawtypes")
-    public <E extends IEvent> void dispatch(E event) {
-        if (this.typeEqual(event.getEventType())) {
-
-            if (LOG.isTraceEnabled()) {
-                LOG.trace("CB size:[{}] cbType:[{}]", this.eventCallback.size(), event.getCallbackType());
-            }
-
-            for (IEventCallback cb : this.eventCallback) {
-                switch (event.getCallbackType()) {
-                case 0:
-                    ((EventCallbackA0) cb).onMiningStarted();
-                    break;
-                case 1:
-                    ((EventCallbackA0) cb).onMiningStopped();
-                    break;
-                case 2:
-                    ((EventCallbackA0) cb).onBlockMiningStarted(event.getFuncArgs().get(0));
-                    break;
-                case 3:
-                    ((EventCallbackA0) cb).onBlockMined(event.getFuncArgs().get(0));
-                    break;
-                case 4:
-                    ((EventCallbackA0) cb).onBlockMiningCanceled(event.getFuncArgs().get(0));
-                    break;
-                default:
-                }
-            }
-        }
     }
 
     /*
